% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GrowthModel}
\alias{GrowthModel}
\title{Growth simulation model}
\usage{
GrowthModel(ln_xdev = NULL, ln_ydev = NULL, obs_err = TRUE,
  tvi_err = TRUE, Pars, Nindiv, ATR_mod)
}
\arguments{
\item{ln_xdev}{are the deviations for each area}

\item{ln_ydev}{are the deviations for each year}

\item{obs_err}{is the observation error}

\item{tvi_err}{is the time-variying individual error}

\item{Pars}{the parameters for the model}
}
\description{
Growth simulation model
}
\author{
Darcy Webber, Jim Thorson
}

