% Generated by roxygen2 (4.0.1): do not edit by hand
\name{normalise}
\alias{normalise}
\title{Function to normalize a vector between two values}
\usage{
normalise(v, a, b)
}
\arguments{
\item{v}{the vector to be normalised}

\item{a}{the lower bound to normalise the vector to}

\item{a}{the upper bound to normalise the vector to}
}
\value{
a vector
}
\description{
Function to normalize a vector between two values
}

