% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_histogram_k.R
\name{plot_histogram_k}
\alias{plot_histogram_k}
\title{Plot histogram of time-varying k}
\usage{
plot_histogram_k(data, report, file_name = "REs_b", delta_t = "weeks")
}
\description{
Plot histogram of time-varying k
}

