% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_histogram_z.R
\name{plot_histogram_z}
\alias{plot_histogram_z}
\title{Plot histogram of time-varying individual growth parameter (z)}
\usage{
plot_histogram_z(data, report, file_name = "REs_z")
}
\description{
Plot histogram of time-varying individual growth parameter (z)
}

