% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_indiv_growth.R
\name{plot_indiv_growth}
\alias{plot_indiv_growth}
\title{Plot individual growth paths}
\usage{
plot_indiv_growth(Sex, Age1, Length1_obs, Length1_hat, Age2, Length2_obs,
  Length2_hat, file_name = "IndivGrowth", file_suffix = "eps")
}
\description{
Plot individual growth paths
}

