% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_indiv_growth_2.R
\name{plot_indiv_growth_2}
\alias{plot_indiv_growth_2}
\title{Plot individual growth paths for two different models}
\usage{
plot_indiv_growth_2(Sex, Age1, Length1_obs, Length1_hat1, Length1_hat2, Age2,
  Length2_obs, Length2_hat1, Length2_hat2, Label = c("No random effects",
  "Transient variation"), file_name = "IndivGrowth", file_suffix = "eps")
}
\description{
Plot individual growth paths for two different models
}

