% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot_obs_pred}
\alias{plot_obs_pred}
\title{Plot observed vs. predicted growth paths}
\usage{
plot_obs_pred(Sex, Length1_obs, Length1_hat, Length2_obs, Length2_hat,
  file_name = "ObsVsPred")
}
\description{
Plot observed vs. predicted growth paths
}

