% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_palette.R
\docType{data}
\name{plot_palette}
\alias{plot_palette}
\title{Colour palette}
\format{\preformatted{ chr [1:8] "#000000" "#E69F00" "#56B4E9" "#009E73" "#F0E442" ...
}}
\usage{
plot_palette
}
\description{
Colour palette used by ggplot2
}
\keyword{datasets}

