% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_simulations.R
\name{plot_simulations}
\alias{plot_simulations}
\title{Function to plot simulations}
\usage{
plot_simulations(directory)
}
\arguments{
\item{directory}{the dir}
}
\value{
plots
}
\description{
Function to plot simulations
}

