% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_standard_resids}
\alias{plot_standard_resids}
\title{Plot standardised residuals}
\usage{
plot_standard_resids(Sex, Length1_obs, Length1_hat, Length2_obs, Length2_hat,
  file_name = "StandardResids")
}
\description{
Plot standardised residuals
}

