% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_all_powers.R
\name{read_all_powers}
\alias{read_all_powers}
\title{Function to read simulations}
\usage{
read_all_powers(directory = "", Nsims = 200)
}
\arguments{
\item{directory}{the directory that the simulations are in}

\item{Nsims}{the number of simulations that were done}
}
\description{
Function to read simulations
}

