% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_all_simulations.R
\name{read_all_simulations}
\alias{read_all_simulations}
\title{Function to read simulations}
\usage{
read_all_simulations(directory = "", Nsims = 200)
}
\arguments{
\item{directory}{the dir}
}
\description{
Function to read simulations
}

