% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_simulations.R
\name{read_simulations}
\alias{read_simulations}
\title{Function to read simulations}
\usage{
read_simulations(directory)
}
\arguments{
\item{directory}{the dir}
}
\description{
Function to read simulations
}

