% Generated by roxygen2 (4.0.2): do not edit by hand
\name{time.step}
\alias{time.step}
\title{Function for converting data.frame from annual to daily/weekly/monthly time steps.}
\usage{
\method{time}{step}(ATR_mod, units = "weeks", year0 = 1972)
}
\arguments{
\item{ATR_mod}{data.frame of data}

\item{units}{choose between days, weeks or months}

\item{year0}{the year that the first fish was born}
}
\value{
a data.frame
}
\description{
Function for converting data.frame from annual to daily/weekly/monthly time steps.
}
\author{
Darcy Webber, Jim Thorson
}

