% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MakeADGrowth.R
\name{MakeADGrowth}
\alias{MakeADGrowth}
\title{Function to create the AD object}
\usage{
MakeADGrowth(data, Options)
}
\arguments{
\item{data}{the data}

\item{Options}{the options}
}
\value{
an AD object
}
\description{
Function to create the AD object
}

