% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_histogram_z}
\alias{plot_histogram_z}
\title{Plot histogram of time-varying individual growth parameter (z)}
\usage{
plot_histogram_z(data, report, file_name = "REs_z")
}
\description{
Plot histogram of time-varying individual growth parameter (z)
}

