% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_parameters.R
\name{plot_parameters}
\alias{plot_parameters}
\title{Pairs plot of parameters}
\usage{
plot_parameters(pars, names, path = ".")
}
\arguments{
\item{pars}{a matrix of the parameters to be plotted}

\item{names}{a vector of the parameter labels}

\item{path}{the dircetory to save the plot in}
}
\value{
a plot
}
\description{
Pairs plot of parameters
}

