% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_standard_resids.R
\name{plot_standard_resids}
\alias{plot_standard_resids}
\title{Plot standardised residuals}
\usage{
plot_standard_resids(Sex, Length1_obs, Length1_hat, Length2_obs, Length2_hat,
  file_name = "StandardResids", file_suffix = "eps")
}
\description{
Plot standardised residuals
}

